+++ src/Lidarr.Api.V1/Config/MediaManagementConfigResource.cs
@@ -53,7 +53,6 @@ namespace Lidarr.Api.V1.Config
                 CopyUsingHardlinks = model.CopyUsingHardlinks,
                 ImportExtraFiles = model.ImportExtraFiles,
                 ExtraFileExtensions = model.ExtraFileExtensions,
-                EnableMediaInfo = model.EnableMediaInfo
             };
         }
     }
