--- src/NzbDrone.Core.Test/MediaFiles/MediaInfo/UpdateMediaInfoServiceFixture.cs
@@ -1,184 +1,0 @@
-using System.IO;
-using FizzWare.NBuilder;
-using Moq;
-using NUnit.Framework;
-using NzbDrone.Common.Disk;
-using NzbDrone.Core.MediaFiles;
-using NzbDrone.Core.MediaFiles.Events;
-using NzbDrone.Core.MediaFiles.MediaInfo;
-using NzbDrone.Core.Test.Framework;
-using NzbDrone.Core.Music;
-using NzbDrone.Test.Common;
-using NzbDrone.Core.Configuration;
-
-namespace NzbDrone.Core.Test.MediaFiles.MediaInfo
-{
-    [TestFixture]
-    public class UpdateMediaInfoServiceFixture : CoreTest<UpdateMediaInfoService>
-    {
-        private Artist _artist;
-
-        [SetUp]
-        public void Setup()
-        {
-            _artist = new Artist
-            {
-                Id = 1,
-                Path = @"C:\artist".AsOsAgnostic()
-            };
-
-            Mocker.GetMock<IConfigService>()
-                  .SetupGet(s => s.EnableMediaInfo)
-                  .Returns(true);
-        }
-
-        private void GivenFileExists()
-        {
-            Mocker.GetMock<IDiskProvider>()
-                  .Setup(v => v.FileExists(It.IsAny<string>()))
-                  .Returns(true);
-        }
-
-        private void GivenSuccessfulScan()
-        {
-            Mocker.GetMock<IVideoFileInfoReader>()
-                  .Setup(v => v.GetMediaInfo(It.IsAny<string>()))
-                  .Returns(new MediaInfoModel());
-        }
-
-        private void GivenFailedScan(string path)
-        {
-            Mocker.GetMock<IVideoFileInfoReader>()
-                  .Setup(v => v.GetMediaInfo(path))
-                  .Returns((MediaInfoModel)null);
-        }
-
-        [Test]
-        public void should_skip_up_to_date_media_info()
-        {
-            var trackFiles = Builder<TrackFile>.CreateListOfSize(3)
-                .All()
-                .With(v => v.RelativePath = "media.flac")
-                .TheFirst(1)
-                .With(v => v.MediaInfo = new MediaInfoModel { SchemaRevision = VideoFileInfoReader.CURRENT_MEDIA_INFO_SCHEMA_REVISION })
-                .BuildList();
-
-            Mocker.GetMock<IMediaFileService>()
-                  .Setup(v => v.GetFilesByArtist(1))
-                  .Returns(trackFiles);
-
-            GivenFileExists();
-            GivenSuccessfulScan();
-
-            Subject.Handle(new ArtistScannedEvent(_artist));
-
-            Mocker.GetMock<IVideoFileInfoReader>()
-                  .Verify(v => v.GetMediaInfo(Path.Combine(_artist.Path, "media.flac")), Times.Exactly(2));
-
-            Mocker.GetMock<IMediaFileService>()
-                  .Verify(v => v.Update(It.IsAny<TrackFile>()), Times.Exactly(2));
-        }
-
-        [Test]
-        public void should_skip_not_yet_date_media_info()
-        {
-            var trackFiles = Builder<TrackFile>.CreateListOfSize(3)
-                .All()
-                .With(v => v.RelativePath = "media.flac")
-                .TheFirst(1)
-                .With(v => v.MediaInfo = new MediaInfoModel { SchemaRevision = VideoFileInfoReader.MINIMUM_MEDIA_INFO_SCHEMA_REVISION })
-                .BuildList();
-
-            Mocker.GetMock<IMediaFileService>()
-                  .Setup(v => v.GetFilesByArtist(1))
-                  .Returns(trackFiles);
-
-            GivenFileExists();
-            GivenSuccessfulScan();
-
-            Subject.Handle(new ArtistScannedEvent(_artist));
-
-            Mocker.GetMock<IVideoFileInfoReader>()
-                  .Verify(v => v.GetMediaInfo(Path.Combine(_artist.Path, "media.flac")), Times.Exactly(2));
-
-            Mocker.GetMock<IMediaFileService>()
-                  .Verify(v => v.Update(It.IsAny<TrackFile>()), Times.Exactly(2));
-        }
-
-        [Test]
-        public void should_update_outdated_media_info()
-        {
-            var trackFiles = Builder<TrackFile>.CreateListOfSize(3)
-                .All()
-                .With(v => v.RelativePath = "media.flac")
-                .TheFirst(1)
-                .With(v => v.MediaInfo = new MediaInfoModel())
-                .BuildList();
-
-            Mocker.GetMock<IMediaFileService>()
-                  .Setup(v => v.GetFilesByArtist(1))
-                  .Returns(trackFiles);
-
-            GivenFileExists();
-            GivenSuccessfulScan();
-
-            Subject.Handle(new ArtistScannedEvent(_artist));
-
-            Mocker.GetMock<IVideoFileInfoReader>()
-                  .Verify(v => v.GetMediaInfo(Path.Combine(_artist.Path, "media.flac")), Times.Exactly(3));
-
-            Mocker.GetMock<IMediaFileService>()
-                  .Verify(v => v.Update(It.IsAny<TrackFile>()), Times.Exactly(3));
-        }
-
-        [Test]
-        public void should_ignore_missing_files()
-        {
-            var trackFiles = Builder<TrackFile>.CreateListOfSize(2)
-                   .All()
-                   .With(v => v.RelativePath = "media.flac")
-                   .BuildList();
-
-            Mocker.GetMock<IMediaFileService>()
-                  .Setup(v => v.GetFilesByArtist(1))
-                  .Returns(trackFiles);
-
-            GivenSuccessfulScan();
-
-            Subject.Handle(new ArtistScannedEvent(_artist));
-
-            Mocker.GetMock<IVideoFileInfoReader>()
-                  .Verify(v => v.GetMediaInfo("media.flac"), Times.Never());
-
-            Mocker.GetMock<IMediaFileService>()
-                  .Verify(v => v.Update(It.IsAny<TrackFile>()), Times.Never());
-        }
-
-        [Test]
-        public void should_continue_after_failure()
-        {
-            var episodeFiles = Builder<TrackFile>.CreateListOfSize(2)
-                   .All()
-                   .With(v => v.RelativePath = "media.flac")
-                   .TheFirst(1)
-                   .With(v => v.RelativePath = "media2.flac")
-                   .BuildList();
-
-            Mocker.GetMock<IMediaFileService>()
-                  .Setup(v => v.GetFilesByArtist(1))
-                  .Returns(episodeFiles);
-
-            GivenFileExists();
-            GivenSuccessfulScan();
-            GivenFailedScan(Path.Combine(_artist.Path, "media2.flac"));
-
-            Subject.Handle(new ArtistScannedEvent(_artist));
-
-            Mocker.GetMock<IVideoFileInfoReader>()
-                  .Verify(v => v.GetMediaInfo(Path.Combine(_artist.Path, "media.flac")), Times.Exactly(1));
-
-            Mocker.GetMock<IMediaFileService>()
-                  .Verify(v => v.Update(It.IsAny<TrackFile>()), Times.Exactly(1));
-        }
-    }
-}
