--- src/NzbDrone.Core/Lidarr.Core.dll.config
@@ -1,7 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<configuration>
-  <dllmap os="osx" dll="MediaInfo.dll" target="libmediainfo.0.dylib"/>
-  <dllmap os="linux" dll="MediaInfo.dll" target="libmediainfo.so.0" />
-  <dllmap os="freebsd" dll="MediaInfo.dll" target="libmediainfo.so.0" />
-  <dllmap os="solaris" dll="MediaInfo.dll" target="libmediainfo.so.0.0.0" />
-</configuration>
