+++ src/NzbDrone.Core/Organizer/FileNameBuilder.cs
@@ -9,7 +9,6 @@ using NzbDrone.Common.Cache;
 using NzbDrone.Common.EnsureThat;
 using NzbDrone.Common.Extensions;
 using NzbDrone.Core.MediaFiles;
-using NzbDrone.Core.MediaFiles.MediaInfo;
 using NzbDrone.Core.Qualities;
 using NzbDrone.Core.Music;
 
@@ -112,7 +111,6 @@ namespace NzbDrone.Core.Organizer
             AddTrackTokens(tokenHandlers, tracks);
             AddTrackFileTokens(tokenHandlers, trackFile);
             AddQualityTokens(tokenHandlers, artist, trackFile);
-            AddMediaInfoTokens(tokenHandlers, trackFile);
 
             var fileName = ReplaceTokens(pattern, tokenHandlers, namingConfig).Trim();
             fileName = FileNameCleanupRegex.Replace(fileName, match => match.Captures[0].Value[0].ToString());
@@ -313,74 +311,6 @@ namespace NzbDrone.Core.Organizer
             tokenHandlers["{Quality Title}"] = m => qualityTitle;
             //tokenHandlers["{Quality Proper}"] = m => qualityProper;
             //tokenHandlers["{Quality Real}"] = m => qualityReal;
-        }
-
-        private void AddMediaInfoTokens(Dictionary<string, Func<TokenMatch, string>> tokenHandlers, TrackFile trackFile)
-        {
-            if (trackFile.MediaInfo == null)
-            {
-                return;
-            }
-
-            var audioCodec = MediaInfoFormatter.FormatAudioCodec(trackFile.MediaInfo);
-            var audioChannels = MediaInfoFormatter.FormatAudioChannels(trackFile.MediaInfo);
-
-            var mediaInfoAudioLanguages = GetLanguagesToken(trackFile.MediaInfo.AudioLanguages);
-            if (!mediaInfoAudioLanguages.IsNullOrWhiteSpace())
-            {
-                mediaInfoAudioLanguages = $"[{mediaInfoAudioLanguages}]";
-            }
-
-            if (mediaInfoAudioLanguages == "[EN]")
-            {
-                mediaInfoAudioLanguages = string.Empty;
-            }
-
-            var mediaInfoSubtitleLanguages = GetLanguagesToken(trackFile.MediaInfo.Subtitles);
-            if (!mediaInfoSubtitleLanguages.IsNullOrWhiteSpace())
-            {
-                mediaInfoSubtitleLanguages = $"[{mediaInfoSubtitleLanguages}]";
-            }
-
-            var videoBitDepth = trackFile.MediaInfo.VideoBitDepth > 0 ? trackFile.MediaInfo.VideoBitDepth.ToString() : string.Empty;
-            var audioChannelsFormatted = audioChannels > 0 ?
-                                audioChannels.ToString("F1", CultureInfo.InvariantCulture) :
-                                string.Empty;
-
-            tokenHandlers["{MediaInfo Audio}"] = m => audioCodec;
-            tokenHandlers["{MediaInfo AudioCodec}"] = m => audioCodec;
-            tokenHandlers["{MediaInfo AudioChannels}"] = m => audioChannelsFormatted;
-
-            tokenHandlers["{MediaInfo Simple}"] = m => $"{audioCodec}";
-
-            tokenHandlers["{MediaInfo Full}"] = m => $"{audioCodec}{mediaInfoAudioLanguages} {mediaInfoSubtitleLanguages}";
-        }
-
-        private string GetLanguagesToken(string mediaInfoLanguages)
-        {
-            List<string> tokens = new List<string>();
-            foreach (var item in mediaInfoLanguages.Split('/'))
-            {
-                if (!string.IsNullOrWhiteSpace(item))
-                    tokens.Add(item.Trim());
-            }
-
-            var cultures = System.Globalization.CultureInfo.GetCultures(System.Globalization.CultureTypes.NeutralCultures);
-            for (int i = 0; i < tokens.Count; i++)
-            {
-                try
-                {
-                    var cultureInfo = cultures.FirstOrDefault(p => p.EnglishName == tokens[i]);
-
-                    if (cultureInfo != null)
-                        tokens[i] = cultureInfo.TwoLetterISOLanguageName.ToUpper();
-                }
-                catch
-                {
-                }
-            }
-
-            return string.Join("+", tokens.Distinct());
         }
 
         private string ReplaceTokens(string pattern, Dictionary<string, Func<TokenMatch, string>> tokenHandlers, NamingConfig namingConfig)
