+++ src/NzbDrone.Core/Parser/Model/MediaInfoModel.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using NzbDrone.Core.Datastore;
+
+namespace NzbDrone.Core.Parser.Model
+{
+    public class MediaInfoModel : IEmbeddedDocument
+    {
+        public string AudioFormat { get; set; }
+        public int AudioBitrate { get; set; }
+        public int AudioChannels { get; set; }
+        public int SchemaRevision { get; set; }
+    }
+}
